# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Task Manager API",
        "contact": {
            "email": "wesleymuniz20@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://mit-license.org"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/projects": {
            "get": {
                "tags": [
                    "Project Endpoint"
                ],
                "summary": "Get list of projects",
                "description": "Returns list paginated of projects",
                "operationId": "8d4a13ef3fc773ff4031e515460919ea",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Query to search projects by title",
                        "schema": {
                            "type": "string",
                            "example": "title"
                        }
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Query to sort by creation: asc|desc",
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "asc": {
                                "summary": "An asc sort --default",
                                "value": "asc"
                            },
                            "desc": {
                                "summary": "An desc sort",
                                "value": "desc"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectCollection"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Project Endpoint"
                ],
                "summary": "Post a new Project",
                "description": "Returns id of the created project",
                "operationId": "a4dd4c17bf5fc24c224357f03dcc848a",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Project created successfully!"
                                                },
                                                "data": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "format": "uuid"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/projects/{id}": {
            "get": {
                "tags": [
                    "Project Endpoint"
                ],
                "summary": "Get project info by id",
                "description": "Returns project data by id",
                "operationId": "af32aed47c19d517816e0b970d60cd5a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Path to find project by id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "2a825edc-95f6-49a9-b46a-d2f296e7e1ce"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Project Endpoint"
                ],
                "summary": "Put project by id",
                "description": "Returns id of the updated project",
                "operationId": "491d11cc6efd20335421a859159e2487",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Path to find project by id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "2a825edc-95f6-49a9-b46a-d2f296e7e1ce"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/ProjectRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Project updated successfully!"
                                                },
                                                "data": {
                                                    "properties": {
                                                        "id": {
                                                            "type": "string",
                                                            "format": "uuid"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Project Endpoint"
                ],
                "summary": "Delete project by id",
                "operationId": "187865e30ff6b72f1cea3037d9dd84fa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Path to find project by id",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "2a825edc-95f6-49a9-b46a-d2f296e7e1ce"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ProjectRequest": {
                "title": "ProjectRequest",
                "properties": {
                    "title": {
                        "type": "string",
                        "maxLength": 100,
                        "minLength": 3
                    }
                },
                "type": "object"
            },
...
            "ProjectCollection": {
                "title": "ProjectCollection",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Project"
                        }
                    }
                },
                "type": "object"
            },
            "ProjectResource": {
                "title": "ProjectResource",
                "properties": {
                    "data": {
                        "$ref": "#/components/schemas/Project"
                    }
                },
                "type": "object"
            },
...
            "Project": {
                "title": "Project",
                "properties": {
                    "id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "user_id": {
                        "type": "string",
                        "format": "uuid"
                    },
                    "title": {
                        "type": "string"
                    },
                    "tasks": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}